remove README contents from main index
Maintaining text in one place is a good idea. Including contents from README is a bit cumbersome to handle and move around text pieces in the index file. Also, multiple places of handling can always be a point of failure / needs redundant maintenance. So, we keep everything in the index file now and the README just mentions the full form of HALO-DROPS and redirects to the documentation hosted at readthedocs