add logging (should close #7)
**Central logging init**

Logging in HALO-DROPS is initialized at a single point: the `__init__` file in the `halodrops` directory within the `src` directory. This means that the first import of any module of the `halodrops` package gets the `halodrops` logger. The idea then is that the modules called from the package should be handled by this logger.

**Logging by modules & entities therein**

We make use of the inbuilt facility of the `logging` module that [any logger with a point in its name technically has a child-parent relationship](https://docs.python.org/3/library/logging.html#logger-objects), i.e. any thing preceding the point is a parent, and that succeeding the point is a child. Therefore, all modules have their logger in the form of `halodrops.hierarchical.path.module`. For example, the `paths` module in the `helper` module has its logger called `halodrops.helper.paths`.

Usually, functions can simply log to their module logger. For classes, however, it is convenient to also have their own logger. Here, we again exploit the parent-child logging relation simply by appending the name of the class at the end. For example, the `Paths` class in the module we referenced above would have a logger `halodrops.helper.paths.Paths`.

**Outputting logs**

The `halodrops` logger and all its children are set to output their logs to three different handlers - two FileHandlers and one StreamHandler. They are differentiated as follows:

| Handler           | Type          | Logging Level | File mode |
| ----------------- | ------------- | ------------- | --------- |
| To file info.log  | FileHandler   | INFO          | append    |
| To file debug.log | FileHandler   | DEBUG         | write     |
| To console        | StreamHandler | WARNING       | write     |

The formatting is the same for all three handlers, i.e.
`"{asctime}  {levelname:^8s} {name:^20s} {filename:^20s} Line:{lineno:03d}:\n{message}"`