add `detect_floater` to Sonde object
For HALO-(AC)3, some sondes seem to have landed on ice. For example, 3 sondes in the 5-sonde circle flown close to the pole (last flight of the campaign, 20220412). Therefore, they kept transmitting data from the surface (outside of their design to sink, which they couldn't in this case). This causes problems with QC.

The function added by this commit should detect if the sonde is a floater.

To determine if a sonde is a floater is difficult with an algorithm because "floaters" behave quite differently. I decided to go for a simple check, i.e. if the gpsalt (altitude from gps) and pressure (pres) has not changed (by more than 1 m and 1 hPa, respectively) in a user-decided number of consecutive timesteps (default 3) for measurements at the surface (decided by gpsalt_threshold, default 25m), then the sonde is a floater.

This is not perfect, but it should work for most cases. The function adds an attribute `is_floater`(True/False) and if that is True, it will also add another attribute `landing_time`. The latter attribute will be used to restrict measurements from the sonde to the time before it landed, for QC and for inclusion in to L2 and onwards.