crop aspen_ds if sonde is a floater
if a sonde is detected as a floater, this method will create another
attr called `cropped_aspen_ds` which will be the same xr.Dataset as
`aspen_ds` except the time coordinates will be cut-off where greater than
the `landing_time` attr.

this is so that later on, QC checks (e.g. profile_fullness) and data levels
from L2 onwards can use the cropped_aspen_ds for floaters.