if floater, cropped_aspen_ds for L2 onwards
if the sonde is a floater, then the data after landing shouldn't
go past QC. Therefore, for L2 and onwards, the dataset would be
ASPEN cropped to landing time, i.e. the `cropped_aspen_ds` attribute
created with the `crop_aspen_ds_to_landing_time` method.

This is so that users do not have to deal with problems that usually
come around with floaters. Of course, it is important to know when the
ASPEN dataset (i.e. L1) has been changed because the sonde was a floater.
Therefore, every L2 file will now carry a global attribute called `is_floater`
which will indicate if the sonde is a floater.